moving the search functionality into its own folder
